Fix the time offset logic.
This addresses #17, the text of which is below.

So there are a couple of different issues here, but they are all fairly tightly related.

First, it's not possible to change the DayNightOffset value through the configuration menu, which is somewhat annoying.  It's also not documented what unit it is in. (seconds as far as I can tell)

Second, the math to apply the timezone offset is a bit screwy.

We take this bit of code:
```
property real timeoffestForDayNight : (Date.now()+(dayNightOffset*1000)+(new Date()).getTimezoneOffset())%86400000/86400000
```

That is, assume that the dayNightOffset is in seconds, and convert that to microseconds, then add the return from getTimezoneOffset as if it is in microseconds, clamp the value to rest within a 24 hour time period, and then divide by the number of microseconds in that 24 hour time period to get a value between 0 and 1.

Except that getTimezoneOffset returns minutes, not microseconds, and the sign is reversed, so UTC-8 is 480 instead of -480 or -28800000.

Looking at the default value of dayNightOffset, my initial guess was that the author lives in GMT+12, but then I realized that the images used for tone and brightness tinting are darkest at the midpoint, and brightest at the start/end, meaning that we want to offset the value by 0.5 so that midnight is the darkest and noon is the brightest.

Which definitely makes dayNightOffset do double duty.

What I ended up with, after fixing a typo or two in variable names which were driving me crazy, was:

```
property real timeoffsetForDayNight : ( Date.now() + (dayNightOffset*1000) + ((new Date()).getTimezoneOffset() * 60 * -1000)) % 864  00000 / 86400000
```