Migrate to Plasma 6
This is just a quick proof of concept to check that this can easily be
ported to Plasma 6. It generates a bunch of warnings and the UI is kind
of messed up but it works.
To make it work, copy the content of the repo to
"/home/pchoisel/.local/share/plasma/wallpapers/com.darkeye.animatedImage/"